Add documentation and reorganize project folder
- Added documentation for:
  - main.go
  - canvas/client.go
  - model/models.go
  - store/config.go
  - store/store.go
  - widgets/navigation.go
  - widgets/screenrouter.go
  - widgets/widgets.go

- Reorganized project folder to better separate documentation and code
- Removed compiled VIO binary
