--- internal/asciiart/months.go
@@ -1,87 +1,0 @@
-package asciiart
-
-func GetMonthHeader(month string) string {
-	if header, ok := MonthHeaders[month]; ok {
-		return header
-	}
-	return ""
-}
-
-// https://patorjk.com/software/taag/#p=display&v=2&f=Bulbhead&t=
-// ----Bulbhead----font---- (old)
-// DiamFont ( NEW CURRENT FONT )
-
-var MonthHeaders = map[string]string{
-	"January": `
-   ▗▖ ▗▄▖ ▗▖  ▗▖▗▖ ▗▖ ▗▄▖ ▗▄▄▖▗▖  ▗▖
-   ▐▌▐▌ ▐▌▐▛▚▖▐▌▐▌ ▐▌▐▌ ▐▌▐▌ ▐▌▝▚▞▘ 
-   ▐▌▐▛▀▜▌▐▌ ▝▜▌▐▌ ▐▌▐▛▀▜▌▐▛▀▚▖ ▐▌  
-▗▄▄▞▘▐▌ ▐▌▐▌  ▐▌▝▚▄▞▘▐▌ ▐▌▐▌ ▐▌ ▐▌  
-                                    `,
-	"February": `
-▗▄▄▄▖▗▄▄▄▖▗▄▄▖ ▗▄▄▖ ▗▖ ▗▖ ▗▄▖ ▗▄▄▖▗▖  ▗▖
-▐▌   ▐▌   ▐▌ ▐▌▐▌ ▐▌▐▌ ▐▌▐▌ ▐▌▐▌ ▐▌▝▚▞▘ 
-▐▛▀▀▘▐▛▀▀▘▐▛▀▚▖▐▛▀▚▖▐▌ ▐▌▐▛▀▜▌▐▛▀▚▖ ▐▌  
-▐▌   ▐▙▄▄▖▐▙▄▞▘▐▌ ▐▌▝▚▄▞▘▐▌ ▐▌▐▌ ▐▌ ▐▌  
-                                        `,
-	"March": `
-▗▖  ▗▖ ▗▄▖ ▗▄▄▖  ▗▄▄▖▗▖ ▗▖
-▐▛▚▞▜▌▐▌ ▐▌▐▌ ▐▌▐▌   ▐▌ ▐▌
-▐▌  ▐▌▐▛▀▜▌▐▛▀▚▖▐▌   ▐▛▀▜▌
-▐▌  ▐▌▐▌ ▐▌▐▌ ▐▌▝▚▄▄▖▐▌ ▐▌
-                          `,
-	"April": `
- ▗▄▖ ▗▄▄▖ ▗▄▄▖ ▗▄▄▄▖▗▖   
-▐▌ ▐▌▐▌ ▐▌▐▌ ▐▌  █  ▐▌   
-▐▛▀▜▌▐▛▀▘ ▐▛▀▚▖  █  ▐▌   
-▐▌ ▐▌▐▌   ▐▌ ▐▌▗▄█▄▖▐▙▄▄▖
-                         `,
-	"May": `
-▗▖  ▗▖ ▗▄▖▗▖  ▗▖
-▐▛▚▞▜▌▐▌ ▐▌▝▚▞▘ 
-▐▌  ▐▌▐▛▀▜▌ ▐▌  
-▐▌  ▐▌▐▌ ▐▌ ▐▌  
-                `,
-	"June": `
-   ▗▖▗▖ ▗▖▗▖  ▗▖▗▄▄▄▖
-   ▐▌▐▌ ▐▌▐▛▚▖▐▌▐▌   
-   ▐▌▐▌ ▐▌▐▌ ▝▜▌▐▛▀▀▘
-▗▄▄▞▘▝▚▄▞▘▐▌  ▐▌▐▙▄▄▖
-                     `,
-	"July": `
-   ▗▖▗▖ ▗▖▗▖ ▗▖  ▗▖
-   ▐▌▐▌ ▐▌▐▌  ▝▚▞▘ 
-   ▐▌▐▌ ▐▌▐▌   ▐▌  
-▗▄▄▞▘▝▚▄▞▘▐▙▄▄▖▐▌  
-                   `,
-	"August": `
- ▗▄▖ ▗▖ ▗▖ ▗▄▄▖▗▖ ▗▖ ▗▄▄▖▗▄▄▄▖
-▐▌ ▐▌▐▌ ▐▌▐▌   ▐▌ ▐▌▐▌     █  
-▐▛▀▜▌▐▌ ▐▌▐▌▝▜▌▐▌ ▐▌ ▝▀▚▖  █  
-▐▌ ▐▌▝▚▄▞▘▝▚▄▞▘▝▚▄▞▘▗▄▄▞▘  █  
-                              `,
-	"September": `
- ▗▄▄▖▗▄▄▄▖▗▄▄▖▗▄▄▄▖▗▄▄▄▖▗▖  ▗▖▗▄▄▖ ▗▄▄▄▖▗▄▄▖ 
-▐▌   ▐▌   ▐▌ ▐▌ █  ▐▌   ▐▛▚▞▜▌▐▌ ▐▌▐▌   ▐▌ ▐▌
- ▝▀▚▖▐▛▀▀▘▐▛▀▘  █  ▐▛▀▀▘▐▌  ▐▌▐▛▀▚▖▐▛▀▀▘▐▛▀▚▖
-▗▄▄▞▘▐▙▄▄▖▐▌    █  ▐▙▄▄▖▐▌  ▐▌▐▙▄▞▘▐▙▄▄▖▐▌ ▐▌
-                                             `,
-	"October": `
- ▗▄▖  ▗▄▄▖▗▄▄▄▖▗▄▖ ▗▄▄▖ ▗▄▄▄▖▗▄▄▖ 
-▐▌ ▐▌▐▌     █ ▐▌ ▐▌▐▌ ▐▌▐▌   ▐▌ ▐▌
-▐▌ ▐▌▐▌     █ ▐▌ ▐▌▐▛▀▚▖▐▛▀▀▘▐▛▀▚▖
-▝▚▄▞▘▝▚▄▄▖  █ ▝▚▄▞▘▐▙▄▞▘▐▙▄▄▖▐▌ ▐▌
-                                  `,
-	"November": `
-▗▖  ▗▖ ▗▄▖ ▗▖  ▗▖▗▄▄▄▖▗▖  ▗▖▗▄▄▖ ▗▄▄▄▖▗▄▄▖ 
-▐▛▚▖▐▌▐▌ ▐▌▐▌  ▐▌▐▌   ▐▛▚▞▜▌▐▌ ▐▌▐▌   ▐▌ ▐▌
-▐▌ ▝▜▌▐▌ ▐▌▐▌  ▐▌▐▛▀▀▘▐▌  ▐▌▐▛▀▚▖▐▛▀▀▘▐▛▀▚▖
-▐▌  ▐▌▝▚▄▞▘ ▝▚▞▘ ▐▙▄▄▖▐▌  ▐▌▐▙▄▞▘▐▙▄▄▖▐▌ ▐▌
-                                           `,
-	"December": `
-▗▄▄▄  ▗▄▄▄▖ ▗▄▄▖▗▄▄▄▖▗▖  ▗▖▗▄▄▖ ▗▄▄▄▖▗▄▄▖ 
-▐▌  █ ▐▌   ▐▌   ▐▌   ▐▛▚▞▜▌▐▌ ▐▌▐▌   ▐▌ ▐▌
-▐▌  █ ▐▛▀▀▘▐▌   ▐▛▀▀▘▐▌  ▐▌▐▛▀▚▖▐▛▀▀▘▐▛▀▚▖
-▐▙▄▄▀ ▐▙▄▄▖▝▚▄▄▖▐▙▄▄▖▐▌  ▐▌▐▙▄▞▘▐▙▄▄▖▐▌ ▐▌
-                                          `,
-}
