--- internal/views/assignments/assignment_page.go
@@ -1,1 +1,0 @@
-package assignments
