--- internal/views/calendar/day_page.go
@@ -1,1 +1,0 @@
-package calendar
