--- internal/views/courses/course_page.go
@@ -1,1 +1,0 @@
-package courses
