--- internal/views/tasks/task_page.go
@@ -1,1 +1,0 @@
-package tasks
