+++ vio-tui/internal/views/assignments/assignment_page.go
@@ -1,0 +1,1 @@
+package assignments
