+++ vio-tui/internal/views/courses/course_page.go
@@ -1,0 +1,1 @@
+package courses
